requirements.txt
joblib
numpy
Pillow
torch
torchvision
